many cleanups
