+++ spec/evm.py
@@ -1,18 +1,6 @@
-from spec import Log, Address, Uint, Hash32, U256
+from eth_types import *
 
 ADD = "\x01"
-
-class Evm:
-    pc: Uint
-    stack: list[U256],
-    memory: bytes,
-    gas_left: Uint
-    current: Address
-    caller: Address
-    data: bytes
-    value: Uint
-    depth: Uint
-    env: EvmEnvironment
 
 class Environment:
     block_hashes: list[Hash32]
@@ -25,10 +13,22 @@ class Environment:
     difficulty: Uint
     state: State
 
-def proccess_call(caller: Address, target: Address, data: bytes, value: Uint, gas: Uint, depth: Uint, env: Environment) -> int, list[Log]
+class Evm:
+    pc: Uint
+    stack: list[U256]
+    memory: bytes
+    gas_left: Uint
+    current: Address
+    caller: Address
+    data: bytes
+    value: Uint
+    depth: Uint
+    env: Environment
+
+def proccess_call(caller: Address, target: Address, data: bytes, value: Uint, gas: Uint, depth: Uint, env: Environment) -> (Uint, list[Log]):
     evm = Evm(
         pc=0,
-        stack: [],
+        stack=[],
         gas_left=gas,
         current=target,
         caller=caller,
@@ -37,11 +37,6 @@ def proccess_call(caller: Address, target: Address, data: bytes, value: Uint, ga
         depth=depth,
         env=env
     )
-
-    env.storage
-    assert ctx.value <= from.balance
-    from.balance -= tx.value
-
 
     code = get_code(evm.env.state, evm.current)
