+++ spec/fixtures/genesis/env.json
@@ -1,0 +1,8 @@
+{
+	"currentCoinbase": "0x0000000000000000000000000000000000000000",
+	"currentDifficulty": "0x400000000",
+	"currentGasLimit": "0x1388",
+	"currentNumber": "0x0",
+	"currentTimestamp": "0x0",
+	"currentExtraData": "0x11bbe8db4e347b4e8c937c1c8370e4b5ed33adb3db69cbdb7a38e1e50b1b82fa"
+}
