+++ spec/fixtures/genesis/txs.json
@@ -1,0 +1,1 @@
+[]
